refactor: match file names to comp
